+++ sharding/consensus/common.go
@@ -5,7 +5,7 @@ import (
 	"github.com/ecoball/go-ecoball/sharding/common"
 )
 
-func (c *Consensus) isVoteEnough(counter uint16) bool {
+func (c *Consensus) isVoteEnough(counter uint32) bool {
 	if counter >= c.ns.GetWorksCounter()*common.DefaultThresholdOfConsensus/1000+1 {
 		return true
 	} else {
